# docs/influxdb_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "历史数据库接口文档",
        "contact": {},
        "version": "1.0 版本"
    },
    "basePath": "/app/influxdb/api/",
    "paths": {
        "/config/influxdb": {
            "get": {
                "description": "查询数据库配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "查询数据库配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-influx_Options"
                        }
                    }
                }
            },
            "post": {
                "description": "修改数据库配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "修改数据库配置",
                "parameters": [
                    {
                        "description": "数据库配置",
                        "name": "cfg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/influx.Options"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-int"
                        }
                    }
                }
            }
        },
        "/query/{pid}/{id}/{name}": {
            "get": {
                "description": "查询历史数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "查询历史数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "变量名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "窗口时间",
                        "name": "window",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "算法",
                        "name": "fn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-array_influx_Point"
                        }
                    }
                }
            }
        },
        "/query/{pid}/{id}/{name}/export": {
            "get": {
                "description": "导出历史数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "导出历史数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "变量名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "窗口时间",
                        "name": "window",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "算法",
                        "name": "fn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-array_influx_Point"
                        }
                    }
                }
            }
        },
        "/rate/{pid}/{id}/{name}": {
            "get": {
                "description": "查询历史数据增长",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rate"
                ],
                "summary": "查询历史数据增长",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "变量名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "窗口时间",
                        "name": "window",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "算法",
                        "name": "fn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-array_influx_Point"
                        }
                    }
                }
            }
        },
        "/rate/{pid}/{id}/{name}/export": {
            "get": {
                "description": "导出历史数据增长",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rate"
                ],
                "summary": "导出历史数据增长",
                "parameters": [
                    {
                        "type": "string",
                        "description": "产品ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "变量名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "窗口时间",
                        "name": "window",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "算法",
                        "name": "fn",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ReplyData-array_influx_Point"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ReplyData-array_influx_Point": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/influx.Point"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-influx_Options": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/influx.Options"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.ReplyData-int": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "influx.Options": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "org": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "influx.Point": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "value": {}
            }
        }
    }
}